Change to int

+++ version_1/Trading-System/src/main/java/TradingSystem/Server/DomainLayer/StoreComponent/Policies/Sales/ProductSale.java
@@ -11,16 +11,16 @@ import java.util.concurrent.ConcurrentHashMap;
 public class ProductSale extends SimpleSale {
 
     Integer productID;
-    Double discountPercentage;
+    Integer discountPercentage;
     TradingSystemImpl tradingSystem = TradingSystemImpl.getInstance();
 
-    public ProductSale(Expression exp, Integer productID, Double discountPercentage) {
+    public ProductSale(Expression exp, Integer productID, Integer discountPercentage) {
         super(exp);
         this.productID = productID;
         this.discountPercentage = discountPercentage;
     }
 
-    public ProductSale(Integer productID, Double discountPercentage) {
+    public ProductSale(Integer productID, Integer discountPercentage) {
         this.productID = productID;
         this.discountPercentage = discountPercentage;
     }
